requirements_ci.txt
fakeRPiGPIO==0.3a0
